+++ Nexus/Changes.md
@@ -1,4 +1,16 @@
 # Utopia Nexus - List of Changes
+## v1.6 (in development)
+### Breaking Changes
+- **RequestHandler** was rewritten to support versioning
+- **ResourceSearchResult.Ready** now takes the ready resource as the first parameter
+  - Also, all **ResourceSearchResult** types now use type parameter C for context.
+### New Features
+- **RequestHandler** now supports versioning (using different resources on different API versions)
+### Other Changes
+- **Request** now contains public constructor parameter `.created` 
+  which holds the creation time of that **Request**
+- **RequestHandler**'s type parameter **C** is now contravariant (**-C**)
+
 ## v1.5.1 - 17.4.2021
 This small update adds utility traits to make Rest resource implementation easier.
 ### New Features